packages/registry-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.11.7
python-dotenv==1.0.0
supabase>=2.24.0
web3==6.15.0
httpx>=0.27.0
PyJWT>=2.8.0
